update:recomment service

+++ src/recomment/recomment.controller.ts
@@ -1,8 +1,12 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, ForbiddenException } from '@nestjs/common';
 import { RecommentService } from './recomment.service';
 import { CreateRecommentDto } from './dtos/create-recomment.dto';
 import { ApiTags } from '@nestjs/swagger';
+import { JwtStrategy } from 'src/auth/guards/jwt.strategy';
+import { LogIn } from 'src/decorators/log-in.decorator';
+import { User } from 'src/user/entities/user.entity';
 
+@UseGuards(JwtStrategy)
 @ApiTags('recomments')
 @Controller('posts')
 export class RecommentController {
@@ -14,8 +18,8 @@ export class RecommentController {
    * @returns
    */
   @Post(':postId/comments/:commentId/recomments')
-  async createRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @Body() createRecommentDto: CreateRecommentDto) {
-    return await this.recommentService.createRecomment(+postId, +commentId, createRecommentDto);
+  async createRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @LogIn() user: User, @Body() createRecommentDto: CreateRecommentDto) {
+    return await this.recommentService.createRecomment(+postId, +commentId, user, createRecommentDto);
   }
 
   /**
@@ -24,7 +28,15 @@ export class RecommentController {
    * @returns
    */
   @Patch(':postId/comments/:commentId/recomments/:recommentId')
-  async updateRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @Param('recommentId') recommentId: number, @Body() createRecommentDto: CreateRecommentDto) {
+  async updateRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @Param('recommentId') recommentId: number, @LogIn() user: User, @Body() createRecommentDto: CreateRecommentDto) {
+    const recomment = await this.recommentService.findRecomment(recommentId);
+
+    if(recomment.user.id != user.id) {
+      throw new ForbiddenException(
+        '작성자가 아닙니다.'
+      )
+    };
+
     return await this.recommentService.updateRecomment(+postId, +commentId, +recommentId, createRecommentDto);
   }
 
@@ -33,7 +45,15 @@ export class RecommentController {
    * @returns
    */
   @Delete(':postId/comments/:commentsId/recomments/:recommentId')
-  async removeRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @Param('recommentId') recommentId: number) {
+  async removeRecomment(@Param('postId') postId: number, @Param('commentId') commentId: number, @Param('recommentId') recommentId: number, @LogIn() user: User, ) {
+    const recomment = await this.recommentService.findRecomment(recommentId);
+
+    if(recomment.user.id != user.id) {
+      throw new ForbiddenException(
+        '작성자가 아닙니다.'
+      )
+    };
+
     return await this.recommentService.removeRecomment(+postId, +commentId, +recommentId);
   }
 }
